requirements.txt
streamlit
sqlalchemy
bcrypt
python-dotenv
stripe